Slår sammen identiske klasser
